refactor(chat): remove ChatRow re-export from chat-commands
