+++ src/chat-message-handler-task-followup.ts
@@ -1,4 +1,4 @@
-import { type ChatRow, createRow } from "./chat-commands";
+import { type ChatRow, createRow } from "./chat-contract";
 import type { Client } from "./client-contract";
 import { t } from "./i18n";
 import { palette } from "./palette";
