+++ src/chat-row-view.tsx
@@ -1,7 +1,7 @@
 import { Box, Text } from "ink";
 import React from "react";
-import type { ChatRow } from "./chat-commands";
 import { renderAssistantContent } from "./chat-content-render";
+import type { ChatRow } from "./chat-contract";
 import { palette } from "./palette";
 import { renderToolOutput as renderToolOutputText, type ToolOutput } from "./tool-output-content";
