+++ src/test-utils.ts
@@ -4,7 +4,8 @@ import { mkdtemp } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import type { AgentMode } from "./agent-contract";
-import type { ChatRow, CommandContext } from "./chat-commands";
+import type { CommandContext } from "./chat-commands";
+import type { ChatRow } from "./chat-contract";
 import type { Message } from "./chat-message-contract";
 import { createMessageHandler } from "./chat-message-handler";
 import { type CreatePickerHandlersInput, createPickerHandlers } from "./chat-picker-handlers";
